Refactors duplicate code into behaviors
