+++ app/forms/hyrax/administrative_report_or_publication_form.rb
@@ -1,8 +1,9 @@
 # Generated via
 #  `rails generate hyrax:work AdministrativeReportOrPublication`
 module Hyrax
-  class AdministrativeReportOrPublicationForm < Hyrax::DefaultWorkForm
-    include ScholarsArchive::NestedGeoBehavior
+  class AdministrativeReportOrPublicationForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
 
     self.model_class = ::AdministrativeReportOrPublication
   end
