+++ app/forms/hyrax/conference_proceedings_or_journal_form.rb
@@ -1,8 +1,10 @@
 # Generated via
 #  `rails generate hyrax:work ConferenceProceedingsOrJournal`
 module Hyrax
-  class ConferenceProceedingsOrJournalForm < Hyrax::ArticleForm
-    include ScholarsArchive::NestedGeoBehavior
+  class ConferenceProceedingsOrJournalForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
+    include ScholarsArchive::ArticleWorkFormBehavior
     self.model_class = ::ConferenceProceedingsOrJournal
   end
 end
