+++ app/forms/hyrax/dataset_form.rb
@@ -1,8 +1,9 @@
 # Generated via
 #  `rails generate hyrax:work Dataset`
 module Hyrax
-  class DatasetForm < Hyrax::DefaultWorkForm
-    include ScholarsArchive::NestedGeoBehavior
+  class DatasetForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
 
     self.model_class = ::Dataset
   end
