+++ app/forms/hyrax/eesc_publication_form.rb
@@ -1,8 +1,9 @@
 # Generated via
 #  `rails generate hyrax:work EescPublication`
 module Hyrax
-  class EescPublicationForm < Hyrax::DefaultWorkForm
-    include ScholarsArchive::NestedGeoBehavior
+  class EescPublicationForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
 
     self.model_class = ::EescPublication
   end
