+++ app/forms/hyrax/graduate_project_form.rb
@@ -1,8 +1,10 @@
 # Generated via
 #  `rails generate hyrax:work GraduateProject`
 module Hyrax
-  class GraduateProjectForm < Hyrax::EtdForm
-    include ScholarsArchive::NestedGeoBehavior
+  class GraduateProjectForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
+    include ScholarsArchive::EtdWorkFormBehavior
     self.model_class = ::GraduateProject
   end
 end
