+++ app/forms/hyrax/graduate_thesis_or_dissertation_form.rb
@@ -1,8 +1,9 @@
 # Generated via
 #  `rails generate hyrax:work GraduateThesisOrDissertation`
 module Hyrax
-  class GraduateThesisOrDissertationForm < Hyrax::EtdForm
-    include ScholarsArchive::NestedGeoBehavior
+  class GraduateThesisOrDissertationForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::EtdWorkFormBehavior
     self.model_class = ::GraduateThesisOrDissertation
   end
 end
