+++ app/forms/hyrax/oer_form.rb
@@ -1,17 +1,10 @@
 # Generated via
 #  `rails generate hyrax:work Oer`
 module Hyrax
-  class OerForm < Hyrax::DefaultWorkForm
-    include ScholarsArchive::DateTermsBehavior
-    include ScholarsArchive::NestedGeoBehavior
+  class OerForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
+    include ScholarsArchive::OerWorkFormBehavior
     self.model_class = ::Oer
-    self.terms += [:resource_type, :is_based_on_url, :interactivity_type, :learning_resource_type, :typical_age_range, :time_required]
-    def primary_terms
-      super
-    end
-
-    def secondary_terms
-      super - self.date_terms + [:is_based_on_url, :interactivity_type, :learning_resource_type, :typical_age_range, :time_required]
-    end
   end
 end
