+++ app/forms/hyrax/open_educational_resource_form.rb
@@ -1,8 +1,10 @@
 # Generated via
 #  `rails generate hyrax:work OpenEducationalResource`
 module Hyrax
-  class OpenEducationalResourceForm < Hyrax::OerForm
-    include ScholarsArchive::NestedGeoBehavior
+  class OpenEducationalResourceForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
+    include ScholarsArchive::OerWorkFormBehavior
     self.model_class = ::OpenEducationalResource
   end
 end
