+++ app/forms/hyrax/technical_report_form.rb
@@ -1,8 +1,9 @@
 # Generated via
 #  `rails generate hyrax:work TechnicalReport`
 module Hyrax
-  class TechnicalReportForm < Hyrax::DefaultWorkForm
-    include ScholarsArchive::NestedGeoBehavior
+  class TechnicalReportForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
     self.model_class = ::TechnicalReport
   end
 end
