+++ app/forms/hyrax/undergraduate_thesis_or_project_form.rb
@@ -1,8 +1,10 @@
 # Generated via
 #  `rails generate hyrax:work UndergraduateThesisOrProject`
 module Hyrax
-  class UndergraduateThesisOrProjectForm < Hyrax::EtdForm
-    include ScholarsArchive::NestedGeoBehavior
+  class UndergraduateThesisOrProjectForm < Hyrax::Forms::WorkForm
+    include ::ScholarsArchive::TriplePoweredProperties::TriplePoweredForm
+    include ScholarsArchive::DefaultWorkFormBehavior
+    include ScholarsArchive::EtdWorkFormBehavior
     self.model_class = ::UndergraduateThesisOrProject
     self.required_fields -= [:degree_level, :degree_name, :degree_field, :degree_grantors, :graduation_year]
